refactor move historic to models as varvaluesmodel ref #61
